feat: Add working ES snapshot repo initialization
