+++ src/backup/s3/cli.py
@@ -4,14 +4,14 @@ from backup.s3 import (
     _upload,
 )
 from backup.options import (
-    dir_options,
+    dir_flags,
 )
 import click
 from pathlib import Path
 
 
 # S3 Flags
-def s3_options(fn):
+def s3_flags(fn):
     options = [
         click.option(
             "--endpoint",
@@ -20,15 +20,6 @@ def s3_options(fn):
             help="S3 endpoint URL",
         ),
         click.option("--bucket", "-b", required=True, help="S3 bucket name"),
-        click.option(
-            "--key", "-k", envvar="ACCESS_KEY", help="S3 access key ID ($ACCESS_KEY)"
-        ),
-        click.option(
-            "--secret",
-            "-s",
-            envvar="SECRET_KEY",
-            help="S3 secret access key ($SECRET_KEY)",
-        ),
     ]
     for option in options:
         fn = option(fn)
@@ -42,15 +33,15 @@ def s3():
 
 
 @s3.command()
-@s3_options
+@s3_flags
 def ls(endpoint: str, bucket: str, key: str, secret: str):
     # TODO: Implement
     pass
 
 
 @s3.command()
-@s3_options
-@dir_options
+@s3_flags
+@dir_flags
 def download(endpoint: str, bucket: str, key: str, secret: str, dir: Path):
     """s3 ➜ local"""
     conf = S3Config(endpoint=endpoint, bucket=bucket, key=key, secret=secret)
@@ -60,8 +51,8 @@ def download(endpoint: str, bucket: str, key: str, secret: str, dir: Path):
 
 
 @s3.command()
-@s3_options
-@dir_options
+@s3_flags
+@dir_flags
 def upload(endpoint: str, bucket: str, key: str, secret: str, dir: Path):
     """local ➜ s3"""
     s3 = S3Config(endpoint=endpoint, bucket=bucket, key=key, secret=secret)
