+++ tests/postgres/test_postgres.py
@@ -1,4 +1,2 @@
-import pytest
-
 def testExample():
     assert True is not False
