+++ tests/s3/test_s3.py
@@ -1,4 +1,2 @@
-import pytest
-
 def testExample():
     assert True is not False
